default.xml: merge rity-nn.xml inside default.xml
Since we officially support nnapi for release v22.0, we merged
meta-nn-extra into meta-nn.

With v22.0 we also want to build all ai/ml stuff in the default image,
so we remove the rity-nn.xml manifest and merge it inside the default
one.

This commit does both at the same time: remove rity-nn.xml and remove meta-nn-extra

Signed-off-by: Julien STEPHAN <jstephan@baylibre.com>
Change-Id: If50affcb8a2bbf54048560d0429f420a6858ad3f